Well i just realised I'm going about this the wrong way.
I neeed to write partial lines to the console.
There is no way of telling if the last line read from a log is partial
or complete, not all logs insert newlines after every message
But with mutliple logs i cant just overwrite the last output line.
So the window writer just got a heck of a lot more complicated.
WAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAH.